# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: API de Servicios Bancarios (Créditos y Cuentas de Ahorro)
  version: 1.2.0
...
servers:
  - url: http://localhost:8082
paths:
...
  /saving/{accountId}/balance: # Nuevo endpoint para consultar saldo de cuenta bancaria
    get:
      summary: Consultar saldo de cuenta bancaria
      description: Endpoint para obtener el saldo actual de una cuenta bancaria.
      tags:
        - BalanceInquiries
      parameters:
        - in: path
          name: accountId
          required: true
          description: Identificador único del producto de cuenta.
          schema:
            type: string
      responses:
        '200': # Respuesta EXITOSA - Saldo encontrado
          description: Saldo de cuenta bancaria obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponseDTO' # DTO para la respuesta de saldo
        '400': # Respuesta de SOLICITUD INVÁLIDA
          description: Solicitud inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO' # DTO de error
        '404': # Respuesta de NO ENCONTRADO
          description: Cuenta bancaria no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO' # DTO de error
...
components:
...
  schemas:
...
    ErrorResponseDTO:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Código de error.
          example: "INVALID_REQUEST"
        message:
          type: string
          description: Mensaje descriptivo del error.
          example: "El campo 'productId' es obligatorio."
...
    BalanceResponseDTO:
      type: object
      required:
        - balance
      properties:
        balance:
          type: number
          format: decimal
          description: Saldo actual disponible.
          example: 1250.75
...